Added VAO, VBO, EBO abstractions
Created abstractions for VAO, VBO and EBO and made use of those implementation in main.c.
Modified Makefile to build and link this abstration in our executable and also changed executable
name to sandbox.out.

Implemented blending and a way to initialize some basic configs. Fragment shader was aslo modified
to facilitate blending.